Begin writing code to compare analyzer and front_end behaviors.
This will help us ensure that the analyzer and the front end interpret
programs in the same way, e.g. they should produce the same type
inference results.

This code is in its own package for now to avoid making the analyzer
package unnecessarily depend on the details of the kernel
representation (which would couple it even more tightly to kernel than
it's already coupled).  We can move it into the analyzer later if we
see fit.

Change-Id: I055e83c7b83b6cd9b1082b4424a2f4144b5abf0e
Reviewed-on: https://dart-review.googlesource.com/72480
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>