Skip non-const expressions when performing one-phase summarization.
The only reason to include non-const expressions in an unlinked
summary is for type inference, but for one-phase summarization, we're
going to perform type inference directly based on the source AST.  So
skip non-const expressions when doing one-phase summarization.

This requires modifying several summary tests so that they no longer
expect the non-const expressions to be present in the unlinked
summary.  It also causes a few tests of one-phase summarization to
start failing, since the AST-based type inference logic hasn't been
implemented yet.  This is ok because one-phase summarization isn't
exposed to customers yet; it will be fixed in follow-up CLs.

Change-Id: I7455fd82b64c59362439206a05a9eb3691c8e397
Reviewed-on: https://dart-review.googlesource.com/73688
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>