Add a doc comment for String.compare
Closes #5612

Similar to `hashCode` and `operator ==`, add an explicit declaration of
the method to attach a doc comment and give extra context beyond the
`Comparable` interface doc comment.

I checked the implementation in `runtime/lib/string_patch.dart` to
validate that it is acting on code units and does not check case.

Change-Id: I85bc94b954a507756132f86cc8299232c73a25cf
Reviewed-on: https://dart-review.googlesource.com/73284
Reviewed-by: Lasse R.H. Nielsen <lrn@google.com>