Replace nestingLevel with a more direct computation of De Bruijn indices.
Not all of the TypeParameterElementImpl constructors were setting it
anyway, and this was causing problems with one-phase summary
generation.  It would have taken a lot of plumbing to make all the
constructors set it correctly; it's easier to just compute it on the
fly when we need it.

Change-Id: I99d2e79c98a8a1dfbfb4e7f1fe90490db779a955
Reviewed-on: https://dart-review.googlesource.com/74672
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>