Handle errors when comparing analyzer and front end behaviors.
For now, if the front end detects an error, we just consider the test
to have passed; we don't try to verify that the analyzer also detects
an error.  This will be improved in a future CL.

Drops the number of failing language_2 tests with `--compiler
compare_analyzer_cfe` from 2676 to 405.

Change-Id: Ib4f3c99710df4a0f0aac9a1c22a9a3279c6a6479
Reviewed-on: https://dart-review.googlesource.com/74968
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>