[vm/compiler] Fix recent regressions due to inaccurate type merging
This is the follow-up fix for
https://github.com/dart-lang/sdk/commit/f81709c3e34438201682237e4faf6cb886094129

As part of that change, merging of known cid and abstract type in
LoadFieldInstr::ComputeType() was done less precisely, loosing a chance
to calculate a more accurate abstract type if cid is known.

This change fixes the problem by resetting abstract type if cid is known.
Similar change is also applied to LoadStaticFieldInstr::ComputeType().

Change-Id: I2b7dfa961e26c0ea98e66aaf9b9afb235dfa5684
Reviewed-on: https://dart-review.googlesource.com/76723
Reviewed-by: Ryan Macnak <rmacnak@google.com>
Commit-Queue: Alexander Markov <alexmarkov@google.com>