[VM] Add regression test for buildbot flakiness due to smi-widening optimization in VM
The added test in this CL causes code to be optimized with _Smi's and
eliminated field guards.  Field guards were eliminated due to usage of Smi ops,
which get removed due to the WidenSmiToInt32 optimization.
A use of the code with _Mint's will then trigger a test failure.

The WidenSmiToInt32 optimization has been disabled for the time being.

Change-Id: I6f7dd003a875b0abd708b3b948cda44378c42198
Reviewed-on: https://dart-review.googlesource.com/c/79540
Reviewed-by: Vyacheslav Egorov <vegorov@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>